+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `table.Session.CopyTables` method
 * Added `x-ydb-trace-id` header into grpc calls
 * Improved topic reader logs
 * Fixed `internal/xstring` package with deprecated warning in `go1.21` about `reflect.{String,Slice}Header`